# openapi.yaml
openapi: 3.0.3
info:
  title: Rancher AI Chat API
  description: API for managing chats and messages with tag-based filtering
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://0.0.0.0:8080
    description: Production server
paths:
...
  /agents/available:
    get:
      summary: List agents available to the current user
      description: >-
        User-scoped. Returns only the agents the requesting user is permitted to
        access, after server-side RBAC filtering by the user's Rancher Global
        Permissions. This is the source of truth for the UI agent selector.
        Fails closed (503) if permissions cannot be resolved.
      operationId: list_available_agents
      tags:
        - Agents
      responses:
        '200':
          description: List of accessible agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableAgent'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Permissions could not be resolved (fail closed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
          description: Error message
      required:
        - message
      example:
        message: userId not found
    AvailableAgent:
      type: object
      description: An agent the current user may access, shaped for the UI selector
      properties:
        name:
          type: string
          description: The AIAgentConfig resource name
        displayName:
          type: string
          description: Human-friendly agent name
        description:
          type: string
          description: Agent description / routing hint
        status:
          type: string
          description: "Readiness derived from the agent's CRD phase: ready or error"
          enum: [ready, error]
      required:
        - name
        - displayName
        - status
      example:
        name: rancher
        displayName: Rancher
        description: Kubernetes cluster operations and Rancher resource management
        status: ready
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: R_SESS
      description: Rancher session cookie for authentication
security:
  - cookieAuth: []
...